+++ work/server/db/models/products.js
@@ -1,7 +1,5 @@
 'use strict';
-const {
-  Model
-} = require('sequelize');
+const { Model } = require('sequelize');
 module.exports = (sequelize, DataTypes) => {
   class Products extends Model {
     /**
@@ -13,32 +11,35 @@ module.exports = (sequelize, DataTypes) => {
       // define association here
     }
   }
-  Products.init({
-    version: DataTypes.INTEGER,
-    density: DataTypes.INTEGER,
-    form: DataTypes.STRING,
-    certificate: DataTypes.STRING,
-    width: DataTypes.INTEGER,
-    lengths: DataTypes.INTEGER,
-    height: DataTypes.INTEGER,
-    tradingMark: DataTypes.STRING,
-    m3: DataTypes.INTEGER,
-    m2: DataTypes.INTEGER,
-    m: DataTypes.INTEGER,
-    widthInArray: DataTypes.INTEGER,
-    m3InArray: DataTypes.INTEGER,
-    densityInDryMax: DataTypes.INTEGER,
-    dinsityInDryDef: DataTypes.INTEGER,
-    humidity: DataTypes.INTEGER,
-    densityHumidityMax: DataTypes.INTEGER,
-    densityHuminityDef: DataTypes.INTEGER,
-    weightMax: DataTypes.INTEGER,
-    weightDef: DataTypes.INTEGER,
-    normOfBrack: DataTypes.INTEGER,
-    coefficientOfFree: DataTypes.INTEGER
-  }, {
-    sequelize,
-    modelName: 'Products',
-  });
+  Products.init(
+    {
+      version: DataTypes.INTEGER,
+      density: DataTypes.INTEGER,
+      form: DataTypes.STRING,
+      certificate: DataTypes.STRING,
+      width: DataTypes.FLOAT,
+      lengths: DataTypes.FLOAT,
+      height: DataTypes.FLOAT,
+      tradingMark: DataTypes.STRING,
+      m3: DataTypes.FLOAT,
+      m2: DataTypes.FLOAT,
+      m: DataTypes.FLOAT,
+      widthInArray: DataTypes.FLOAT,
+      m3InArray: DataTypes.FLOAT,
+      densityInDryMax: DataTypes.FLOAT,
+      dinsityInDryDef: DataTypes.FLOAT,
+      humidity: DataTypes.FLOAT,
+      densityHumidityMax: DataTypes.FLOAT,
+      densityHuminityDef: DataTypes.FLOAT,
+      weightMax: DataTypes.FLOAT,
+      weightDef: DataTypes.FLOAT,
+      normOfBrack: DataTypes.INTEGER,
+      coefficientOfFree: DataTypes.INTEGER,
+    },
+    {
+      sequelize,
+      modelName: 'Products',
+    }
+  );
   return Products;
 };
